Update hive.js
port collisions bad.  The docs say that the number of ports is so high you should never hit it unless something is wrong.  Lets put that to the test.  (Most likely some kind of port reset is going to be necessary, but after the world explodes I'll know when that needs to happen.)